LayoutTests/imported/mathml-in-html5/mathml/presentation-markup/operators/mo-axis-height-1.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>mo axis height</title>
<link rel="help" href="http://www.mathml-association.org/MathMLinHTML5/S3.html#SS2.SSS4">
<meta name="assert" content="Element mo correctly uses the axis height parameter from the MATH table.">
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<style>
  math, mspace {
    font-size: 10px;
  }
  @font-face {
    font-family: axisheight5000-verticalarrow14000;
    src: url("../../../fonts/math/axisheight5000-verticalarrow14000.woff");
  }
</style>
<script>
  var emToPx = 10 / 1000; // font-size: 10px, font.em = 1000
  var epsilon = 1;

  function getBox(aId) {
    return document.getElementById(aId).getBoundingClientRect();
  }

  setup({ explicit_done: true });
  window.addEventListener("load", function() {
    document.fonts.ready.then(function() {
      window.setTimeout(runTests, 250);
    });
  });

  function runTests() {
    test(function() {
      var v1 = 5000 * emToPx;
      var moMiddle = (getBox("mo").bottom + getBox("mo").top) / 2;
      assert_approx_equals(getBox("baseline").bottom - moMiddle,
                           v1, epsilon, "mo: axis height");
    }, "AxisHeight");

    done();
  }
</script>
</head>
<body>
  <p>
    <math style="font-family: axisheight5000-verticalarrow14000;">
      <mrow>
        <mspace id="baseline" mathbackground="blue" width="100px" height="1px"/>
        <mo id="mo" mathcolor="green">&#x21A8;</mo>
        <mspace mathbackground="black" width="10px" height="100px"/>
      </mrow>
    </math>
  </p>
</body>
</html>
