LayoutTests/imported/mathml-in-html5/mathml/presentation-markup/operators/mo-axis-height-1-expected.txt

↨


FAIL AxisHeight assert_approx_equals: mo: axis height expected 50 +/- 1 but got 70
